Move buildStores from d2-stores to d2-store-factory
